+++ src/main/java/lotto/service/LottoService.java
@@ -3,23 +3,23 @@ package lotto.service;
 import lotto.domain.Lotto;
 import lotto.dto.LottoDto;
 import lotto.utils.RandomNumberGenerator;
-import lotto.utils.Validator;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import static lotto.validator.LottoServiceValidator.validatePurchaseAmount;
+
+
 public class LottoService {
 
     private final static int LOTTO_PRICE = 1000;
-    private final Validator validator;
 
-    public LottoService(Validator validator) {
-        this.validator = validator;
+    public LottoService() {
     }
 
     public List<LottoDto> buyLottos(int purchaseAmount) {
-        validator.validatePurchaseAmount(purchaseAmount);
+        validatePurchaseAmount(purchaseAmount);
 
         int count = purchaseAmount / LOTTO_PRICE;
         List<LottoDto> purchasedLotto = new ArrayList<>();
