+++ src/main/java/lotto/validator/LottoServiceValidator.java
@@ -1,0 +1,10 @@
+package lotto.validator;
+
+public class LottoServiceValidator {
+
+    public static void validatePurchaseAmount(int amount) {
+        if (amount % 1000 != 0) {
+            throw new IllegalArgumentException("[ERROR] 구입 금액은 1000원 단위여야 합니다.");
+        }
+    }
+}
